refactor: LottoMachine 클래스 삭제에 따른 NumberConstant 생성
LottoMachine 클래스 삭제에 따른 NumberConstant 생성

- LottoMachine 클래스 내 상수 import 모두 NumberConstant로 변경
