+++ src/main/java/lotto/common/constant/NumberConstant.java
@@ -1,0 +1,9 @@
+package lotto.common.constant;
+
+public final class NumberConstant {
+    public static final int LOTTO_SIZE = 6;
+    public static final int MINIMUM_LOTTO_RANGE = 1;
+    public static final int MAXIMUM_LOTTO_RANGE = 45;
+
+    private NumberConstant() {}
+}
