+++ src/main/java/lotto/common/exception/ErrorMessage.java
@@ -1,0 +1,4 @@
+package lotto.common.exception;
+
+public enum ErrorMessage {
+}
